refactor: split index search into lineairdb_index_search
Move the index search planning and execution layer out of ha_lineairdb.cc into lineairdb_index_search.{cc,hh}, renaming index_search_plan.hh to the new header and relocating build_search_plan, the execute_* operations, and the order-by range-scan helpers as a pure relocation with no behavior change.
