+++ proxy/CMakeLists.txt
@@ -44,6 +44,7 @@ set(LINEAIRDB_SOURCES ha_lineairdb.cc ha_lineairdb.hh
                       lineairdb_proxy.cc lineairdb_proxy.hh
                       lineairdb_keyenc.cc lineairdb_keyenc.hh
                       lineairdb_pushdown.cc lineairdb_pushdown.hh
+                      lineairdb_index_search.cc lineairdb_index_search.hh
                       rpc_trace.cc rpc_trace.hh)
 add_definitions(-DMYSQL_SERVER)
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-error -Wextra -mcx16 -fPIC")
